Image Classification using Resnet50 and Keras.
It also deploys uses flask for the backend and javascript for parsing the information. Resnet50 belongs to the module python keras (keras.applications.resnet50). 'preprocess_input' is used to process the image through a list of information and dataset already present in keras. 'decode_predictions' is used to decode the informatino recieved from the now processed image.